refactor: reorganise the code structure into separate files for the accuracy metric and Neural Network creation
