+++ linalgzero/models/simple_cnn.py
